Single point for ldd, male ELF.libs and SSH.libs return consistent output of {full local path: ldd address}
